requirements.txt
click==8.0.1
commonmark==0.9.1
cryptography==3.4.7
future==0.18.2
protobuf==3.17.3
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
pymongo==3.12.0
opentracing>=2.0,<2.4
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
dnspython==1.16.0